Change of Yellow Warning to Red warning
Change of Yellow Warning to Red warning for the catch TCPKeepAlive equal
to 0, which means we didn't find it in the registry or it actually
equals this. Determine this change makes it easier to see the difference
in the catch for customers and SEs to determine what is actually going
on.