Bug Fix, Minor Upgrade
-Export-Clixml provided an error "String was not recognized as a valid
DateTime", to what I assume is because of the language/time format
change after the install occurs. Because this is a slient error in this
cmdlet, I moved the check for errors before this cmdlet. Now we should
see this error any longer.
-Unable to get power plan setting. Not sure of the conditions this can
occur, but it was reported by Apoc70. No harm should come to include
this error handling.
-Included the friendly name of the NIC in the output. Upgrade recommend
by Apoc70.