Added packets received discarded counter
Added the ability to see if we are dropping packets at the NIC which can cause a lot of issues.

Currently it should work in most scenarios, but i think there might be an issue with older OS versions that might not provide the correct adapter description so we can't match that adapter to the correct performance counter. However, this should be rare to see.

Resolved #55