buildRevision improvement
This pull request will fix issue #130 . This pull request replaces pull #131 .
We use invariant culture to convert FileBuildPart and FilePrivatePart to double.
This is an better way instead of using a function to switch the culture on the thread.

Invariant culture:
An invariant culture is culture-insensitive. Your application specifies the invariant culture by name using an empty string ("") or by its identifier. InvariantCulture defines an instance of the invariant culture. It is associated with the English language but not with any country/region. It is used in almost any method in the Globalization namespace that requires a culture.