Visual C++ Redistributable Package Check
First implementation of an Visual C++ Redistributable Package check to fulfill issue #134 .
It's a little bit complex to do this check nice and clear. I was not able to find a full version overview incl. version information for Visual C++ Redistributable 2012 and 2013. 
I've got the latest version from here: https://support.microsoft.com/en-us/help/2977003/the-latest-supported-visual-c-downloads.
We validate only if the latest known version of the required packaged (based on Exchange server role) is installed and if not, we show a red warning dialog with download resource to the latest version. We do this check for Exchange 2013 +

As this is the first implementation, there's space to improve this check.