Updated Checking for Freespace on drives
Some of the data that is collected can be extremely large, because of this drive space has always been a concern. The biggest threat this script has is filing up the OS drive and causing the server to stop working.

With this huge change, we are checking for the available free space before we copy over every single item to make sure that we don't run into this issue. If we don't have enough free space, it doesn't copy over that data.

Currently it requires 10GB of free space to start and to keep that 10GB cushion the entire time, otherwise we start to fail to copy items.