Query Windows Update Catalog for replacing updates
We query the Windows Update Catalog to detect replacing updates to address issue #237 .  If we found problematic July 2019 updates and no fixing updates, we're going to query the Windows Update Catalog for any replacing updates that supersed the initial fixing update. We do this by using the newly introduced Get-ReplacingServerUpdates function. It'll work only if we're connected to the internet.