Updated Confirm-ExchangeShell to accommodate for the errors that can occur
In Confirm-ExchangeShell, we call a script that can have errors occur within it. We don't see them on the screen, however, this throws off the script's logic to find errors that can occur and report them.

Added in logic to account for this. In doing so, we updated the Invoke-Actions method as well.