Added logic to include the errors that occur in Export-Clixml
Added the same logic that is in place Confirm-ExchangeShell to find all the errors that occurs while loading the management shell. This should be the best way to keep the InstalledOn property in the Get-Hotfix command & keep the current logic in place to make sure we export the data prior to checking to see if we have any issues with the script to keep the warning at the end of the script.

Addressed #288