Vulnerability Check improvements (#285)
Goal:
Reduce complexity and amount of code

We can now pass multiple builds when using Test-VulnerabilitiesByBuildNumbersandDisplay method.
We have to call the method just for the latest CU for which we've published a specific security update (SU). For example: If there is an SU for Exchange 2019 CU3 and CU4, we have to call the method in CU4 if block, passing SecurityFixedBuild numbers for CU3 as well as CU4 (make sure that the build numbers are in ascending order).