Added new analyzer engine with display to screen function
The newest analyzer function that is going to handle both the display to screen results as well as the HTML options that we wish to display.

This will also include an object that you can use to do pester testing against for a feature later down the road to do automatic testing.

The HTML piece isn't fully written at this time.