Improved Save-WindowsEventLogs
Instead of provided the full name for each log that we want to get copied, let's just copy out the the "category" that we want. This way if any new logs are created, this shouldn't need to be updated, only new categories.

Because we are doing this list based off what is in the logging directory, we don't need to have a loop to validate the files.