Added in a script logger that logs down everything
Now able to log everything that the scirpt is doing and in each of the instances. Only problem right now is i wont be able to capture the zipping process of the enter folder.

Might need to make some tweaking once I start working on the functionality of being able to run this script from a tools box.