Changed order of disabling extra
When doing a DiskShadow backup, we don't aways truncate the logs right away. We want to wait inside of the of optin to remove or leave the backup that occurred to allow extra to continue collecting information of the truncation.

Don't need to do this in 3rd party backup as we can just wait till the enter occurs.