Change into the output path location for diskshadow
If we don't change into the output path location, we get the following error message:

"The .cab metadata file cannot be stored in the current working directory, because it is read-only"

and the script just continue without there seeming like an issue. Quickly change into the output location and pop back to the current working directory to avoid any issues like this.