Added check for Microsoft Exchange Writer
Microsoft Exchange Writer needs to be present when trying to do a backup. If it isn't, we aren't going to be able to properly backup the database and truncate the logs.

Fail the script because it isn't going to be able to provide any useful information as to why the databases aren't trucating logs if that writer isn't present.