Added CollectOverMetrics.ps1 script collection option
Added the ability to run the CollectOverMetrics.ps1 script on the servers. We are able to only do this if the local machine that is executing the script is an Exchange Server or a tools box. Otherwise, we aren't able to run this as we can't invoke the script inside of a Invoke-Command.

Resolved #22