Added the ability to load a config file
This allows you to be able to provide a config file at the same location and have it run with those settings. This way you don't need to provide the parameters every time you want to run the script. You just execute the script it finds the config file and load up the variables and runs with the script.