Added ErrorAction on Get-ServerMonitoringOverride
On my CAS only server getting an error when trying to run this. However, the old version didn't have a CASOnly check here. Not sure if this is something new or something with CASOnly servers now.

Either way, we should attempt to collect the data if we can. Added ErrorAction to silently continue to handle this.