Removed display of warning of SecurityProtocol not matching registry settings
In a few different labs, not able to get this setting to change regardless of what I set in the registry. However, Lukas is able to get a different result. There must be another check that we need to do.

Need to make a new release, want to include what we have thus far but remove the warning for the time being.