Save results by default
First, this change actually fixes #96. We missed a spot.

This change also alters the syntax. There are three ways to
run the script:

This works as before:

`.\Test-ProxyLogon.ps1 -OutPath`

This sets -OutPath to $PSScriptRoot\Test-ProxyLogonLogs:

`.\Test-ProxyLogon.ps1`

This only displays results and does not save them:

`.\Test-ProxyLogon.ps1 -DisplayOnly`