Fix single hit problem for 26855 test
PR #184 introduced a single hit bug on the 26855 test. Prior to this
PR, Get-Cve26855 returned a collection, and because we wrapped the
call in @(), it was still a collection even if only one object was
returned. With this PR, Get-Cve26855 now returns a PSCustomObject,
and that means wrapping the result in @() actually _causes_ a
single hit problem.