Multiple files found with same RevisionNumber
Addressed issue where we find multiple files with the same revision number and don't do anything about it. We need to log it better and also test out if they are the same hash value. If they are, it is the same file and it is safe to copy it over.

Fixed #418