Remove CompareExchangeHashes script
This script is no longer recommended or supported, so it can be
removed. Anyone who absolutely must use the old script for some reason
can download the final release:

https://github.com/microsoft/CSS-Exchange/releases/download/v21.04.21.1229/CompareExchangeHashes.ps1

That release uses the download location of the final hash file release,
rather than looking at the latest release, so it will continue working
as long as the release is there.