Tie up some loose ends
Abandon work on fixing things other than bad permissions. It's
making things very complicated. Maybe we can revisit that in
the future. For now, we're back to -RemoveInvalidPermissions to fix
the one thing we can. Otherwise, we're outputting basically the same
instructions as before on how to fix other problems.