Addressed Build script to avoid excluding references of possible dot loading within itself
It is possible to have comment and or warnings where we have two periods in the same line and it looks like we are dot loading the file. When we just have a comment or a warning providing what to do within the same file, don't exclude it from the list that we want to publish.