Added Identity to Confirm-ExchangeShell
To avoid returning a large list of Exchange server for testing Exchange Management Shell, adding in an Identity parameter to pass.

It is possible that the identity is not an Exchange Server, thus causing this to fail as if EMS is not loaded. This may cause misleading information in the end result. If we see an increase of question on the script(s) regarding this, will need to add in some logic to address it at that time.