Fixed Pester testing not liking $true
When passing to a function of $true to test against, we aren't comparing a boolean to a boolean. it is also marked true if a string has a value against a true. If you make a "True" compare on the test, then we get the correct results we are looking for.