Fixed issue with Win32_PhysicalMemory returning a null value
Had this as a critical data collection, however, we only really need the memory from it to do calculations and we can get a close value from the in32_ComputerSystem which is what we used before.

Decided to revert back to a noncritical and if null, use the in32_ComputerSystem Total Physical Memory instead.