If on tools server determine if Identity needs to be removed
Need to determine if we are on a tools server or not. In some cases, we need to bypass the filter of the Identity of Get-ExchangeServer. For example when using the load balancer switch we need to do this.

To still allow the script to fail if we don't have a proper Exchange Server identity in some cases, added a parameter of IgnoreToolsIdentity. We should only set this to true in the cases we need it.