Open Relay Wild Card Domain Issue
If there is an Open Relay, we should report it as an error as they shouldn't be having that set. Customers should be using a receive connector for this. There was an update with code that wasn't anticipating this that is causing a crash to occur. We want to call this out if this is set for that as well.