Reworked the TLS settings collection and storing
To be able to have more advanced logic in the data collection, like to be able to tell if something is misconfigured, needed to improve the logic within the Get-AllTlsSettingsFromRegistry. Because of the name and the fact that we are wanting to also collect the Get-TlsCipherSuite as well, created a new function called Get-AllTlsSettings, which then calls the Get-AllTlsSettingsFromRegistry and removes the SecurityProtocol call from the registry function.