Remove additional wording
This is not required to provide.

None is not a requirement to be considered secure if mitigation is in place.

This wording could come up if someone uses IP Filtering to block other IPs and not have the correct EP configured on the virtual directory. Causing this wording to appear, which does not line up to the issue.