Remove permissions without piping
Fixes #1881

Once upon a time, we had to pipe invalid permissions to get them
to remove. Testing in my lab on current versions of Exchange
indicates this is no longer the case, so I'm removing the pipe.
This will allow permissions removal to work in Remote PowerShell.

If we find that piping is still necessary for some scenarios, we
may have to revisit this change. This will depend on user
feedback.